Add MyNinjaTeam
