+++ exceptions-and-errors-lab/MyNinjaTeam/app_exceptions/DivisionByZeroException.java
@@ -1,0 +1,18 @@
+package app_exceptions;
+
+/**
+ * @author Trespeces, Guillermo II O.
+ * @willj067
+ */
+
+// checked exception - must be declared or handled
+public class DivisionByZeroException extends Exception {
+
+    public DivisionByZeroException() {
+        super("Cannot divide by zero.");
+    }
+
+    public DivisionByZeroException(String message) {
+        super(message);
+    }
+}
