+++ exceptions-and-errors-lab/MyNinjaTeam/app_exceptions/NegativeNumberException.java
@@ -1,18 +1,7 @@
 package app_exceptions;
 
-/**
- * @author Trespeces, Guillermo II O.
- * @willj067
- */
-
-// unchecked exception - extends RuntimeException
-public class InvalidInputException extends RuntimeException {
-
-    public InvalidInputException() {
-        super("Invalid input provided.");
-    }
-
-    public InvalidInputException(String message) {
+public class NegativeNumberException extends Exception {
+    public NegativeNumberException(String message) {
         super(message);
     }
 }
